fix: prefer PDF when other PDF has URL
When searching for the preferred PDF, Zotero first checks if the URL
field of the PDF attachment matches the parent item, then looks at
dateAdded. But somehow, if another PDF has URL and the selected item
does not, even though the URL from that PDF does not match the parent
item, Zotero still thinks it's the prefered PDF. As a workaround, I have
to set the URL field of the selected PDF explicitly.